requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
websockets==11.0.3
pillow==10.0.0
pyautogui==0.9.54
pynput==1.7.6
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
numpy==1.24.3
gunicorn==21.2.0
python-dotenv==1.0.0
